Rename files to support OSX case insensitive FS
